feat(security+perf+ops): Phase 39 — sprint d'urgence post-redeploy
5 fixes critiques après l'audit complet d'aujourd'hui.

## 🔴 Sécurité (data leaks)

1. **POST /api/photos/spoof** : sans aucun `Depends(require_user)`.
   N'importe qui pouvait spoofer photos sans cookie ni crédit débité.
   Fix : ajout du dépendant.

2. **Cross-tenant via tokens upload** : les tokens d'upload (UUID
   random sans owner) étaient devinables/interceptables. User A
   pouvait utiliser le token de user B dans son /api/render/batch
   pour render avec sa vidéo.
   Fix : préfixe `u{user.id}_` au token côté /api/render/upload,
   validation au create_batch ET create_spoof_batch. Si mismatch
   → 400 "Token d'upload invalide (pas le tien ou expiré)".

## 🟠 Bug live (crédits)

3. **render_credits int → float dans tous les Pydantic schemas**.
   La colonne DB est `Float` depuis Phase 38 (0.5 crédit/spoof)
   mais `UserOut.render_credits: int` et `UserSummary` aussi
   tronquaient silencieusement → un user à 12.5 voyait 12 partout
   (sidebar, admin table). UI pensait avoir moins de crédits.
   Fix : `float` côté UserOut + UserSummary + UserCreate + UserUpdate.

## 🟠 Performance

4. **Indexes DB manquants**. Aucun index sur les colonnes filter/sort
   utilisées partout. Scan full table à chaque list templates / jobs
   / tags. OK avec 50 rows, désastreux à 1000+.
   Fix : `index=True` sur Template.owner_id, RenderJob.owner_id +
   status, et migration `CREATE INDEX IF NOT EXISTS` au boot pour
   les DB existantes (idempotent).
   Aussi : `render_workers=1 → 2` (config) pour paralléliser 2
   renders en même temps, et `PRAGMA busy_timeout=5000` pour éviter
   les SQLITE_BUSY 500 random sous charge.

## 🟠 Ops

5. **Aucun backup SQLite jusqu'ici**. Si la VM crash on perdait tout
   (cas vécu aujourd'hui). Nouveau script `scripts/backup_db.ps1` :
   - sqlite3 .backup vers C:\Botmontage\backups\botmontage-{date}.db
   - rotation auto sur 7 jours
   - fallback Copy-Item si sqlite3 pas dans le PATH
   - à scheduler dans Windows Task Scheduler (instructions dans le
     header du script).
   Bonus : ajouter rclone vers Drive/S3 dans le bloc final si tu
   veux du offsite.

Frontend : UserMeSchema doc à jour pour préciser que render_credits
peut être float (z.number() acceptait déjà).

+++ backend/app/api/admin.py
@@ -54,7 +54,10 @@ class UserSummary(BaseModel):
     role: UserRole
     priority: UserPriority
     max_templates: Optional[int]
-    render_credits: int
+    # Phase 39 — Float depuis Phase 38 (0.5 crédit/spoof). Avant : int
+    # qui tronquait silencieusement les .5 → admin voyait "12 crédits"
+    # alors que la DB en avait 12.5.
+    render_credits: float
     is_active: bool
     # extra computed counts for the admin table view
     template_count: int = 0
@@ -68,7 +71,7 @@ class UserCreate(BaseModel):
     priority: UserPriority = UserPriority.normal
     # null = unlimited. Default 5 for new regular users.
     max_templates: Optional[int] = Field(default=5, ge=0)
-    render_credits: int = Field(default=50, ge=0)
+    render_credits: float = Field(default=50.0, ge=0)
 
 
 class UserUpdate(BaseModel):
@@ -79,7 +82,7 @@ class UserUpdate(BaseModel):
     role: Optional[UserRole] = None
     priority: Optional[UserPriority] = None
     max_templates: Optional[int] = Field(default=None, ge=0)
-    render_credits: Optional[int] = Field(default=None, ge=0)
+    render_credits: Optional[float] = Field(default=None, ge=0)
     is_active: Optional[bool] = None
